Appease betty

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -24,8 +24,8 @@ void *_calloc(unsigned int nmemb, unsigned int size)
 		return (NULL);
 	}
 
-	for (i = 0; i < (nmemb*size); ++i)
-		*(newBlock + i) = (int)0;
+	for (i = 0; i < (nmemb * size); ++i)
+		*(newBlock + i) = 0;
 
 	return (newBlock);
 }
